Made setSat function for saturation properties setSat_pX and setSat_TX, that includes dew and bubble properties. In C wrapper all sat variables are accompanied with sat in the variable name..
Made 2 choices on two-phase partial derivatives. Note that the numerical ones uses only one additional state, and that the more correct method would be to use 2 points, but increases CPU time. Also the analytical ones at constant concentration are pseudo, so they do not include the fact that liquid and vapor properties also depend on concentration. This should however be more quick.

Made a choice on calculation of transport properties.